notebooklm-audio-overview/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "notebooklm-audio-overview"
version = "0.1.0"
description = "Generate and download NotebookLM Audio Overviews from link lists with structured JSON output."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "notebooklm-py[browser]>=0.3.4",
]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Operating System :: OS Independent",
]

[project.optional-dependencies]
dev = [
  "pytest>=9.0.2",
]

[project.scripts]
notebooklm-audio-overview = "skills.notebooklm_audio_overview.cli:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["skills*"]
namespaces = true

[tool.setuptools.package-data]
"skills.notebooklm_audio_overview" = ["examples/*.json"]
